Refactor index.tsx for clarity
